put old ration code (from new NRC) into different folder which is not used (added readme). replaced all rations and nutrient requirements with hardcoded values. changed usage of feed instance so it runs without error
