+++ Inputs/animal_management.json
@@ -479,7 +479,7 @@
             "feed_quality_table": "feed_quality",
             "nutrient_table": "nutrients",
 
-            "purchased_feeds": [9, 34, 46, 92],
+            "purchased_feeds": [2, 24, 34, 38, 91, 102, 137],
             "growing_feeds": [8, 87],
 
             "storage_options": {
